feat: refactor profile screen regenerative game details widget
