+++ lib/screens/profile_screen/profile_screen.dart
@@ -7,6 +7,7 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:intl/intl.dart';
 import 'package:loading_overlay/loading_overlay.dart';
 import 'package:ootopia_app/screens/home/components/home_store.dart';
+import 'package:ootopia_app/screens/profile_screen/components/regenerative_game_details.dart';
 import 'package:ootopia_app/screens/profile_screen/components/wallet_bar_widget.dart';
 import 'package:ootopia_app/screens/wallet/wallet_screen.dart';
 import 'package:ootopia_app/screens/wallet/wallet_store.dart';
@@ -227,181 +228,17 @@ class _ProfileScreenState extends State<ProfileScreen> {
                             fontWeight: FontWeight.w500),
                       ),
                       SizedBox(height: 4),
-                      Container(
-                        height: 46,
-                        decoration: BoxDecoration(
-                            border: Border.fromBorderSide(BorderSide(
-                                width: 1,
-                                color: Color(0xff101010).withOpacity(.1))),
-                            borderRadius: BorderRadius.circular(45)),
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 24),
-                          child: Row(
-                            mainAxisSize: MainAxisSize.min,
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              InkWell(
-                                onTap: () {
-                                  showModalBottomSheet(
-                                    context: context,
-                                    barrierColor: Colors.black.withAlpha(1),
-                                    backgroundColor: Colors.black.withAlpha(1),
-                                    builder: (BuildContext context) {
-                                      return SnackBarWidget(
-                                        menu: AppLocalizations.of(context)!
-                                            .regenerationGame,
-                                        text: AppLocalizations.of(context)!
-                                            .theRegenerationGame,
-                                        buttons: [
-                                          ButtonSnackBar(
-                                            text: AppLocalizations.of(context)!
-                                                .learnMore,
-                                            onTapAbout: () {
-                                              Navigator.of(context)
-                                                  .pushNamedAndRemoveUntil(
-                                                PageRoute.Page.homeScreen.route,
-                                                (Route<dynamic> route) => false,
-                                                arguments: {
-                                                  "returnToPageWithArgs": {
-                                                    'currentPageName':
-                                                        "learning_tracks"
-                                                  }
-                                                },
-                                              );
-                                            },
-                                          )
-                                        ],
-                                        marginBottom: true,
-                                      );
-                                    },
-                                  );
-                                },
-                                child: RichText(
-                                  text: TextSpan(children: [
-                                    TextSpan(
-                                      text: AppLocalizations.of(context)!
-                                              .personalGoal +
-                                          ": ",
-                                      style: TextStyle(
-                                          fontSize: 16, color: Colors.black87),
-                                    ),
-                                    TextSpan(
-                                      text:
-                                          "${authStore.currentUser?.dailyLearningGoalInMinutes ?? 00}min",
-                                      style: TextStyle(
-                                          fontSize: 16,
-                                          color: Colors.black87,
-                                          fontWeight: FontWeight.bold),
-                                    ),
-                                  ]),
-                                ),
-                              ),
-                              SizedBox(
-                                width: 8,
-                              ),
-                              Text("|",
-                                  style: TextStyle(
-                                      fontSize: 16,
-                                      color: Colors.black87,
-                                      fontWeight: FontWeight.bold)),
-                              SizedBox(
-                                width: 8,
-                              ),
-                              Row(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  InkWell(
-                                    onTap: () {
-                                      showModalBottomSheet(
-                                        context: context,
-                                        barrierColor: Colors.black.withAlpha(1),
-                                        backgroundColor:
-                                            Colors.black.withAlpha(1),
-                                        builder: (BuildContext context) {
-                                          return SnackBarWidget(
-                                              menu:
-                                                  AppLocalizations.of(context)!
-                                                      .laurelWreath,
-                                              text: AppLocalizations.of(
-                                                      context)!
-                                                  .laurelWreathRepresentHowManyTimesAPersonHasReachedTheirGoalInTheRegenerationGame,
-                                              buttons: [
-                                                ButtonSnackBar(
-                                                  text: AppLocalizations.of(
-                                                          context)!
-                                                      .learnMore,
-                                                  onTapAbout: () {
-                                                    Navigator.of(context)
-                                                        .pushNamedAndRemoveUntil(
-                                                      PageRoute.Page.homeScreen
-                                                          .route,
-                                                      (Route<dynamic> route) =>
-                                                          false,
-                                                      arguments: {
-                                                        "returnToPageWithArgs":
-                                                            {
-                                                          'currentPageName':
-                                                              "learning_tracks"
-                                                        }
-                                                      },
-                                                    );
-                                                  },
-                                                )
-                                              ],
-                                              marginBottom: true);
-                                        },
-                                      );
-                                    },
-                                    child: Row(
-                                      children: [
-                                        SvgPicture.asset(
-                                          "assets/icons_profile/laurel_wreath.svg",
-                                          width: 24,
-                                          height: 21,
-                                          color: Color(0xff018f9c),
-                                        ),
-                                        SizedBox(
-                                          width: 8,
-                                        ),
-                                        Text(
-                                            "${store?.profile!.totalTrophyQuantity!}",
-                                            style: TextStyle(
-                                              fontSize: 17,
-                                              fontWeight: FontWeight.bold,
-                                              color: Color(0xff018f9c),
-                                            )),
-                                      ],
-                                    ),
-                                  ),
-                                  SizedBox(
-                                    width: 8,
-                                  ),
-                                  InkWell(
-                                    onTap: () {
-                                      setState(() {
-                                        if (isVisible) {
-                                          isVisible = false;
-                                        } else {
-                                          isVisible = true;
-                                        }
-                                      });
-                                    },
-                                    child: RotationTransition(
-                                      turns: !isVisible
-                                          ? AlwaysStoppedAnimation(270 / 360)
-                                          : AlwaysStoppedAnimation(90 / 360),
-                                      child: Icon(
-                                        Icons.arrow_back_ios_rounded,
-                                        color: Color(0xff03145C),
-                                        size: 12,
-                                      ),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ],
-                          ),
-                        ),
+                      RegenerativeGameDetails(
+                        isVisible: isVisible,
+                        onArrowTap: () {
+                          setState(() {
+                            if (isVisible) {
+                              isVisible = false;
+                            } else {
+                              isVisible = true;
+                            }
+                          });
+                        },
                       ),
                       SizedBox(
                         height: GlobalConstants.of(context).spacingNormal,
